♻️[Refactor] 채팅 리펙토링 #283 (#284)
* 뒤로가기 버튼 추가  #245

* 채팅방 헤더 누르면 해당 모임 상세 페이지로 이동하기 구현 #245

* 메시지 입력 부분 화면 하단에 고정 #245

* 채팅 입력창 화면 너비에 맞게 수정 #283

* refreshAccessToken 분리 #283
